src/requirements.txt
asgiref==3.5.2
Django==4.1.4
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.14.0
psycopg2-binary==2.9.5
pytz==2022.7
sqlparse==0.4.3
